requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.1.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.61.0
fastapi-utils==0.2.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
orjson==3.3.1
packaging==20.4
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.0.1
pytest-env==0.6.2
python-dotenv==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
toml==0.10.1
ujson==3.1.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
zipp==3.1.0
